requirements.txt
Django==1.11.9
django-modeltranslation==0.12.2
Pillow==5.2.0
pkg-resources==0.0.0
pytz==2018.5
